+++ src/app/globals.css
@@ -1,6 +1,13 @@
 @import url("./reset.css");
 @import url("./variables.css");
 
+.root_container {
+  max-width: 500px;
+  height: 100vh;
+  margin: 0 auto;
+  position: relative;
+}
+
 @font-face {
   font-family: "Pretendard";
   src: url("https://cdn.jsdelivr.net/gh/Project-Noonnu/noonfonts_2107@1.1/Pretendard-Regular.woff") format("woff");
