+++ src/app/page.tsx
@@ -1,5 +1,6 @@
 import MainLayout from "./_main/mainLayout";
+import Login from "./login/page";
 
 export default function Home() {
-  return <MainLayout />;
+  return <Login />;
 }
